scripts/requirements.txt
boto3==1.34.0
faker==25.9.1
python-dotenv==1.0.1
